feat(c.rocks): Add admins -> suspended user DMs
Also works for non-suspended and remote users,
but effectively just creates a DM room
between server admins and the target user.
Suspended users are given an explicit bypass to send
events in this room, and are forbidden from leaving it.

This also implicitly allows them to DM admins in a pre-existing DM room,
but cannot create nor join new ones due to being suspended.

Unsuspended users are still able to leave admin DMs.

# Conflicts:
#	src/admin/user/mod.rs